[PAGOPA-2978] fix: defining payment amounts as decimal values

+++ fdr/src/main/scala/eu/sia/pagopa/common/json/model/rendicontazione/Flow.scala
@@ -46,7 +46,7 @@ case class Flow(
                  revision: Option[Int],
                  status: Option[String],
                  computedTotPayments: Integer,
-                 computedSumPayments: Integer,
+                 computedSumPayments: Double,
                  regulationDate: String,
                  regulation: String,
                  sender: Sender,
